Use separate variables instead of overloading incorrect usage of high.xrecoff
Advance xlog pointer to the next segment when partial file is not found. This
bug was masked by the fact that we stored a segment id in xrecoff instead of
an offset...

Finally, change the test for file-found to use ||, otherwise we would miss the
first segment in each new logid, if pg_receivexlog was shut down while at
segment 0, and the log advanced past segment 0 before it was restarted.